# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Timefly API
  description: MongoDB ExpressJS NodeJS REST API for a Project Management App.<br />Github repository available [here](https://github.com/aneguet/MEVN_ProjectManagementApp).<br />Some endpoints require `auth-token` in order to be executed.
  version: 1.0.0
servers:
  - url: https://timefly-mevn.herokuapp.com/api/
    description: Heroku remote server
  - url: http://localhost:4000/api/
    description: Localhost
components:
  securitySchemes:
    ApiKeyAuth: # arbitrary name for the security scheme
      type: apiKey
      in: header # can be "header", "query" or "cookie"
      name: auth-token
  schemas:
    User:
      type: object
      required:
        - 'first_name'
        - 'last_name'
        - 'email'
        - 'password'
      properties:
        role:
          type: string
          default: 'user'
          enum: [user, admin]
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
          default: '123456'
        avatar:
          type: string
          default: 'https://i.imgur.com/e5YmXMB.png'
          description: 'Must be a direct link to a hosted image'
        registration_date:
          type: string
          description: 'It takes the current date as default'
          example: '2014-01-01T23:28:56.782Z'
        weekly_hours:
          type: number
          default: 0
...
paths:
  /users/register:
    post:
      summary: Register new User
      tags:
        - User
      description: ''
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'

      responses:
        201:
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  data:
                    type: string
        400:
          description: Client error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /users/login:
    post:
      summary: Login user in order to get JWT token (auth-token)
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        200:
          description: User successfully logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                      user:
                        $ref: '#/components/schemas/User'

        400:
          description: Client error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /users:
    get:
      summary: Get all users (Admin functionality)
      tags:
        - User
      security:
        - ApiKeyAuth: []
      responses:
        200:
          description: Users successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        400:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        401:
          description: Access denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        403:
          description: This request requires an Admin role
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /users/userLogin:
    get:
      summary: Get user by the logged in user ID (For Login ONLY)
      tags:
        - User
      security:
        - ApiKeyAuth: []
      responses:
        200:
          description: Login user successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        400:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        401:
          description: Access denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /users/user:
    get:
      summary: Get user by ID  (Admin functionality)
      tags:
        - User
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: header
          description: User ID
          required: true
          type: string
      responses:
        200:
          description: User successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        400:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        401:
          description: Access denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        403:
          description: This request requires an Admin role
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    put:
      summary: Update user by ID
      tags:
        - User
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      parameters:
        - name: id
          in: header
          description: User ID
          required: true
          type: string
      responses:
        200:
          description: User successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        400:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        401:
          description: Access denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        404:
          description: The user you want to update does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    delete:
      summary: Delete user by ID
      tags:
        - User
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: header
          description: User ID
          required: true
          type: string
      responses:
        201:
          description: User successfully deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        400:
          description: Token is not valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        401:
          description: Access denied
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        404:
          description: The user you want to delete does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...